requirements_hf.txt
gradio==4.44.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
easyocr==1.7.0
scipy==1.11.3